use common validateOwnership func in cert controller
